Always specify versionRange for dependencies in mods.toml
The default versionRange value used by FML is "", which formally means that no versions can satisfy the dependency. For some reason, this still usually works most of the time, but fails on dedicated servers. Manually specify a range of "*" to match any version. Fix #612. This issue does not affect NeoForge.